pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spectrum-os"
version = "0.1.0"
description = "光谱 OS — SQLite-backed multi-agent knowledge pipeline"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy[asyncio]>=2.0.30",
    "aiosqlite>=0.20.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "typer>=0.13.0",
    "tenacity>=9.0.0",
    "httpx>=0.27.0",
    "tavily-python>=0.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[project.scripts]
spectrum = "spectrum.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/spectrum"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
